dew-cis-part1/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-order==1.2.0
requests==2.31.0
httpx==0.27.0